chore(pluggable-widgets-tools): switch back to jest
