+++ packages-common/pluggable-widgets-tools/package.json
@@ -104,15 +104,14 @@
   },
   "devDependencies": {
     "@mendix/generator-widget": "^8.9.0",
-    "async": "^3.2.0",
-    "async-mutex": "^0.2.2",
+    "cross-env": "^5.2.0",
     "fs-extra": "9.0.0",
     "shelljs": "^0.8.4",
     "tree-kill": "^1.2.2",
     "yeoman-test": "^2.6.0"
   },
   "scripts": {
-    "test": "npm run test:cross-platform -- --limited",
-    "test:cross-platform": "node tests/commands.js"
+    "test": "cross-env LIMIT_TESTS=true npm run test:cross-platform",
+    "test:cross-platform": "jest --testTimeout=500000 tests/commands.spec.js"
   }
 }
